Move search menu to other component
