+++ src/features/AdminPage/StreetcodesTable/StreetcodesTable.component.tsx
@@ -1,14 +1,14 @@
-import './StreetcodesTable.styles.scss';
+import './SearchMenu.styles.scss';
 import StreetcodesApi from "@/app/api/streetcode/streetcodes.api";
 import { useAsync } from "@/app/common/hooks/stateful/useAsync.hook";
 import Streetcode, { Status } from "@/models/streetcode/streetcode-types.model";
 import Table from "antd/es/table/Table";
 import { useEffect, useState } from "react";
-import MagnifyingGlass from '@images/header/Magnifying_glass.svg';
 import { Button, Input, Select, SelectProps } from 'antd';
 import { DeleteOutlined, FormOutlined, RollbackOutlined } from '@ant-design/icons';
 import FRONTEND_ROUTES from '@/app/common/constants/frontend-routes.constants';
 import useMobx from '@/app/stores/root-store';
+import SearchMenu from './SearchMenu.component';
 
 const StreetcodesTable = () => {
 
@@ -132,44 +132,10 @@ const StreetcodesTable = () => {
         
     }, [streetcodes]); 
 
-    const options: SelectProps['options'] = [
-        { value: 'published', label: 'опублікований' },
-        { value: 'draft', label: 'чернетка' },
-        { value : 'deleted', label: 'видалений' }
-      ];
-
-    const handleChange = (value: string) => {
-    console.log(`selected ${value}`);
-    };
-
     return(
     <>
         <div className="StreetcodeTableWrapper">
-            <div className='searchMenu'>
-            <div className='searchMenuElement'>
-                <Button className='Button'>Пошук стріткодів</Button>
-            </div>
-            <div className='searchMenuElement'>
-                <Input
-                    className='searchMenuElementInput'
-                    prefix={<MagnifyingGlass />}
-                    placeholder="Назва або індекс"
-                />
-            </div>
-                <div className='searchMenuElement'>
-                    <Select
-                        mode="multiple"
-                        allowClear
-                        className='searchMenuStatusSelected'
-                        placeholder="Статус стріткодів"
-                        onChange={handleChange}
-                        options={options}
-                    />
-                </div>
-                <div className='searchMenuElement'>
-                    <Button className='Button' onClick={() => window.open(`${FRONTEND_ROUTES.STREETCODE.BASE}/new-streetcode`,'_blank')}>Новий стріткод</Button>
-                </div>
-            </div> 
+            <SearchMenu/> 
                 <Table columns={columnsNames}
                 dataSource={mapedStreetCodes}
                 pagination={{className: "paginationButton", pageSize: 8}}
